+++ lib/holepuncher.js
@@ -12,7 +12,7 @@ const NAT_UNKNOWN = Symbol.for('NAT_UNKNOWN')
 const NAT_PORT_RANDOMIZED = Symbol.for('NAT_PORT_RANDOMIZED')
 
 module.exports = class Holepuncher {
-  constructor (address, onconnected) {
+  constructor (address) {
     const self = this
 
     this.address = address
@@ -23,7 +23,6 @@ module.exports = class Holepuncher {
     this.localFriendly = false
     this.remoteFriendly = false
     this.lan = false
-    this.onconnected = onconnected || noop
 
     this._onmessage = onmessage
     this._allSockets = []
@@ -40,13 +39,20 @@ module.exports = class Holepuncher {
     this._onconnection = onconnection
 
     function onconnection (rawSocket) {
+      if (self._resolveConnection === null) {
+        self._stop(null)
+        rawSocket.on('error', noop)
+        rawSocket.destroy()
+        return
+      }
+
       const socket = this
 
       socket.setTTL(DEFAULT_TTL)
       socket.firewall(true)
       self._stop(null)
-
-      self.onconnected(rawSocket)
+      self._resolveConnection(rawSocket)
+      self._resolveConnection = null
     }
 
     function onmessage (buf, rinfo) {
@@ -185,7 +191,6 @@ module.exports = class Holepuncher {
     // if (this.address.type === NAT_OPEN && !this._isClient) return
 
     this._bindAllSockets()
-    if (this._stopped) return
 
     if (!this.localFriendly && !this.remoteFriendly && !this.lan) return
