+++ lib/kat-session-store.js
