+++ lib/messages.js
@@ -4,7 +4,7 @@ const { peerIPv4, peerIPv4Array } = require('dht-rpc/lib/messages') // TODO: mov
 
 exports.peerIPv4 = peerIPv4
 
-exports.katSession = {
+exports.announceSelf = {
   preencode (state, m) {
     state.end += 32 + 64
   },
@@ -20,7 +20,7 @@ exports.katSession = {
   }
 }
 
-exports.katConnectRelay = {
+exports.connectRelay = {
   preencode (state, m) {
     cenc.buffer.preencode(state, m.noise)
     state.end += 32 + 2
@@ -39,7 +39,7 @@ exports.katConnectRelay = {
   }
 }
 
-exports.katConnect = {
+exports.connect = {
   preencode (state, m) {
     cenc.buffer.preencode(state, m.noise)
     state.end += 32
