+++ lib/noise.js
