+++ lib/timer.js
